SECURITY FIX: Use empty fragment countermeasure on SSL 3.0.
TLS Lite was using the empty fragment countermeasure to prevent
chosen plaintext attacks (like BEAST), however for some reason
this countermeasure was only applied to TLS 1.0, and not
SSL 3.0 (TLS 1.1 and higher are immune).

Anyways, this was probably just a typo, so the protection is
extended to SSL 3.0.  However, this countermeasure is known
to be incompatible with some sites, so we should investigate
the 1/n-1 countermeasure.