Oops further fix to RFC 5054 SRP support.
The "missing username" idiom was previously implemented as a
non-fatal alert, and all handshake messages were hashed together.

In RFC 5054, the "missing username" idiom is indicated by a
fatal unknown_psk_identity alert, which prompts the client
to restart a whole knew handshake.

In particular, the hash state needs to be reset prior to this
second handshake.