Pulled out all the callback / handshakeUnknown stuff from client.
The missing SRP username idiom assumed the socket would stay
open, which is not a good assumption.  It's better to have the
caller of the handshake function handle the idiom (prompt for
username/pass, re-call handshake function); instead of the
handshake function.

The cert callback was still sorta useful, since there's
no error thrown or anything on CertificateRequest,  we just
ignore it if no client certChain was passed in.

I don't care that much about Client Cert authentication.
But, maybe a mechanism should br added to get an exception
from CertificateRequest.  Anyways, not an immediate priority,
and I'd rather simplify the code.