Factored out RSA keyexchange from server handshake.
Some associated changes:

 - The code was following some weird post-handshake way of
delivering certain error messages, due to the fact that the
client would only handle error message if it encountered
them during a read; and if a socket.error occurred while the
client was still sending, the client would not notice the
error message that had arrived in the read buffer.

So, I changed it so that if a socket.error occurs during
sending of a handshake message, the code reads the next
record to see if it's an alert that might explain why the
connection was suddenly closed.

 - Began the process of explicitly setting numBytes(..., howMany)
in numBytes, the absence of which was probably an intermittent bug,
so I need to search through the code for this.