use FIFO for integration tests synchronisation
because Travis-CI servers are sometimes under heavy load, the
synchronisation using `time.sleep(2)` is not deterministic,
change it to use Unix FIFO

Server writes a single character (that's non-blocking) to FIFO
when it has set up and is able to handle requests from client.
The client tries to read a single character (that's blocking)
before trying to connect to server.