fix all problems with tests
this is a squashed merge of pull requests #72, #75 and #76

update the expired certificate

The previous certificate had very short validity period, create a
new one that has 10 year lifetime

use socket for integration tests synchronisation

because Travis-CI servers are sometimes under heavy load, the
synchronisation using `time.sleep(2)` is not deterministic,
change it to use extra socket for synchronisation

Server writes a single character (that's non-blocking) to connection
when it has set up and is able to handle requests from client.
The client tries to read a single character (that's blocking)
before trying to connect to server.

test with tackpy

update the TACKs to the new certificate so that we can use them for
testing

also add instruction to the Makefile on how to generate them
from scratch

add tests with tackpy installed to the travis test matrix