Fix python2 compatibility issue with X509 DER parsing
The documentation for X509.parseBinary claims to support python2 str as an input.
The input string is correctly converted to bytearray, but the array is not
passed to the ASN1Parser - the parser gets the original string and
fails with Type error when attempting to logical-or the characters with an integer.